Created the DoubleOutputRedirection,InputRedirection, and OutputRedirection classes
- Made modification on cmdTranslator to suport the new
classes ; doubleoutput redirection, Input redirection ,
and output redirection
- In some cases,the following classes can be combined with other
connectors such as And,Or ,semicolon.
- Parentheses can apply in the command.
- Pipe has been modified to work with redirection classes
- Although the redirection classes work with connectors
and parenthese, there are other cases , it won't work.

/InputRedirection/
- Input redirection execute function:
Assigns the input file descriptor to open the file
 to read only flag  and the right child will get the
filename from the argument list index 0.
- Then the left child will execute the input file descriptor and
output file descriptor

/OutputRedirection/
- Output redirection execute function:
Assigns the output file descriptor to open the file with
flags to write only, check if the file exist , and
 if the previous file exists.
- A few modes were passed in the open function to restrict the
user to do certain tasks in the file only.
- Then the left child will execute the input file descriptor and
outputfile descriptor

/DoubleOuputRedirection/
- DoubleoutputRedirection execute function:
Assigns the double file descriptor to open the file to write to
- The first paramter ; the right child will get the filname from
the argument list index 0
- The second parameter restrict the user to append and write only
- The third paramter will restrict the user from writing crazy stuff
in the file descriptor